# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.JSONError:
    properties:
      error:
        type: string
    type: object
  models.CreateUserRequest:
    properties:
      email:
        example: test@user.com
        type: string
      first_name:
        example: test
        type: string
      last_name:
        example: user
        type: string
      password:
        example: SecurePass123!
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  models.ListOfUsersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.UserResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  models.LoginRequest:
    properties:
      email:
        example: test@user.com
        type: string
      password:
        example: SecurePass123!
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      rating:
        type: integer
      role:
        type: string
      updated_at:
        type: string
    type: object
  models.VoteRequest:
    properties:
      value:
        enum:
        - -1
        - 0
        - 1
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  title: User Management
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login and get JWT token
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/handlers.JSONError'
      summary: User Login
      tags:
      - users
  /users:
    get:
      consumes:
      - application/json
      description: Get list of all users from DB
      parameters:
      - description: Limit records per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: 'Sort order: asc or desc (default desc)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListOfUsersResponse'
        "500":
          description: Failed to get users
          schema:
            $ref: '#/definitions/handlers.JSONError'
      summary: Get list of users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User registration info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid input or validation info
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.JSONError'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user by ID from DB
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "403":
          description: You can delete only your own account
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "500":
          description: Failed to delete user
          schema:
            $ref: '#/definitions/handlers.JSONError'
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get User By ID from DB
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.JSONError'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information by its UUID
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: User update info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "500":
          description: Failed to update user
          schema:
            $ref: '#/definitions/handlers.JSONError'
      security:
      - BearerAuth: []
      summary: Update user by ID
      tags:
      - users
  /users/{id}/vote:
    post:
      consumes:
      - application/json
      description: 'Vote +1 or -1 for a user profile, or 0 to withdraw. Limit: 1 vote
        per minute.'
      parameters:
      - description: Target User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Vote value (-1, 0, 1)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.VoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Self voting not allowed / Invalid input
          schema:
            $ref: '#/definitions/handlers.JSONError'
        "429":
          description: Rate limit exceeded
          schema:
            $ref: '#/definitions/handlers.JSONError'
      security:
      - BearerAuth: []
      summary: Vote for a user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"